# implement a memory-optimized time.Time replacement