--- a/conn_id_generator.go
+++ b/conn_id_generator.go
@@ -5,6 +5,7 @@ import (
 	"slices"
 	"time"
 
+	"github.com/quic-go/quic-go/internal/monotime"
 	"github.com/quic-go/quic-go/internal/protocol"
 	"github.com/quic-go/quic-go/internal/qerr"
 	"github.com/quic-go/quic-go/internal/wire"
@@ -38,7 +39,7 @@ func (cr connRunners) ReplaceWithClosed(ids []protocol.ConnectionID, b []byte, e
 }
 
 type connIDToRetire struct {
-	t      time.Time
+	t      monotime.Time
 	connID protocol.ConnectionID
 }
 
@@ -95,7 +96,7 @@ func (m *connIDGenerator) SetMaxActiveConnIDs(limit uint64) error {
 	return nil
 }
 
-func (m *connIDGenerator) Retire(seq uint64, sentWithDestConnID protocol.ConnectionID, expiry time.Time) error {
+func (m *connIDGenerator) Retire(seq uint64, sentWithDestConnID protocol.ConnectionID, expiry monotime.Time) error {
 	if seq > m.highestSeq {
 		return &qerr.TransportError{
 			ErrorCode:    qerr.ProtocolViolation,
@@ -123,7 +124,7 @@ func (m *connIDGenerator) Retire(seq uint64, sentWithDestConnID protocol.Connect
 	return m.issueNewConnID()
 }
 
-func (m *connIDGenerator) queueConnIDForRetiring(connID protocol.ConnectionID, expiry time.Time) {
+func (m *connIDGenerator) queueConnIDForRetiring(connID protocol.ConnectionID, expiry monotime.Time) {
 	idx := slices.IndexFunc(m.connIDsToRetire, func(c connIDToRetire) bool {
 		return c.t.After(expiry)
 	})
@@ -149,21 +150,21 @@ func (m *connIDGenerator) issueNewConnID() error {
 	return nil
 }
 
-func (m *connIDGenerator) SetHandshakeComplete(connIDExpiry time.Time) {
+func (m *connIDGenerator) SetHandshakeComplete(connIDExpiry monotime.Time) {
 	if m.initialClientDestConnID != nil {
 		m.queueConnIDForRetiring(*m.initialClientDestConnID, connIDExpiry)
 		m.initialClientDestConnID = nil
 	}
 }
 
-func (m *connIDGenerator) NextRetireTime() time.Time {
+func (m *connIDGenerator) NextRetireTime() monotime.Time {
 	if len(m.connIDsToRetire) == 0 {
-		return time.Time{}
+		return 0
 	}
 	return m.connIDsToRetire[0].t
 }
 
-func (m *connIDGenerator) RemoveRetiredConnIDs(now time.Time) {
+func (m *connIDGenerator) RemoveRetiredConnIDs(now monotime.Time) {
 	if len(m.connIDsToRetire) == 0 {
 		return
 	}