--- a/internal/ackhandler/interfaces.go
+++ b/internal/ackhandler/interfaces.go
@@ -1,8 +1,7 @@
 package ackhandler
 
 import (
-	"time"
-
+	"github.com/quic-go/quic-go/internal/monotime"
 	"github.com/quic-go/quic-go/internal/protocol"
 	"github.com/quic-go/quic-go/internal/wire"
 )
@@ -10,19 +9,19 @@ import (
 // SentPacketHandler handles ACKs received for outgoing packets
 type SentPacketHandler interface {
 	// SentPacket may modify the packet
-	SentPacket(t time.Time, pn, largestAcked protocol.PacketNumber, streamFrames []StreamFrame, frames []Frame, encLevel protocol.EncryptionLevel, ecn protocol.ECN, size protocol.ByteCount, isPathMTUProbePacket, isPathProbePacket bool)
+	SentPacket(t monotime.Time, pn, largestAcked protocol.PacketNumber, streamFrames []StreamFrame, frames []Frame, encLevel protocol.EncryptionLevel, ecn protocol.ECN, size protocol.ByteCount, isPathMTUProbePacket, isPathProbePacket bool)
 	// ReceivedAck processes an ACK frame.
 	// It does not store a copy of the frame.
-	ReceivedAck(f *wire.AckFrame, encLevel protocol.EncryptionLevel, rcvTime time.Time) (bool /* 1-RTT packet acked */, error)
-	ReceivedBytes(_ protocol.ByteCount, rcvTime time.Time)
-	DropPackets(_ protocol.EncryptionLevel, rcvTime time.Time)
-	ResetForRetry(rcvTime time.Time)
+	ReceivedAck(f *wire.AckFrame, encLevel protocol.EncryptionLevel, rcvTime monotime.Time) (bool /* 1-RTT packet acked */, error)
+	ReceivedBytes(_ protocol.ByteCount, rcvTime monotime.Time)
+	DropPackets(_ protocol.EncryptionLevel, rcvTime monotime.Time)
+	ResetForRetry(rcvTime monotime.Time)
 
 	// The SendMode determines if and what kind of packets can be sent.
-	SendMode(now time.Time) SendMode
+	SendMode(now monotime.Time) SendMode
 	// TimeUntilSend is the time when the next packet should be sent.
 	// It is used for pacing packets.
-	TimeUntilSend() time.Time
+	TimeUntilSend() monotime.Time
 	SetMaxDatagramSize(count protocol.ByteCount)
 
 	// only to be called once the handshake is complete
@@ -32,23 +31,23 @@ type SentPacketHandler interface {
 	PeekPacketNumber(protocol.EncryptionLevel) (protocol.PacketNumber, protocol.PacketNumberLen)
 	PopPacketNumber(protocol.EncryptionLevel) protocol.PacketNumber
 
-	GetLossDetectionTimeout() time.Time
-	OnLossDetectionTimeout(now time.Time) error
+	GetLossDetectionTimeout() monotime.Time
+	OnLossDetectionTimeout(now monotime.Time) error
 
-	MigratedPath(now time.Time, initialMaxPacketSize protocol.ByteCount)
+	MigratedPath(now monotime.Time, initialMaxPacketSize protocol.ByteCount)
 }
 
 type sentPacketTracker interface {
 	GetLowestPacketNotConfirmedAcked() protocol.PacketNumber
-	ReceivedPacket(_ protocol.EncryptionLevel, rcvTime time.Time)
+	ReceivedPacket(_ protocol.EncryptionLevel, rcvTime monotime.Time)
 }
 
 // ReceivedPacketHandler handles ACKs needed to send for incoming packets
 type ReceivedPacketHandler interface {
 	IsPotentiallyDuplicate(protocol.PacketNumber, protocol.EncryptionLevel) bool
-	ReceivedPacket(pn protocol.PacketNumber, ecn protocol.ECN, encLevel protocol.EncryptionLevel, rcvTime time.Time, ackEliciting bool) error
+	ReceivedPacket(pn protocol.PacketNumber, ecn protocol.ECN, encLevel protocol.EncryptionLevel, rcvTime monotime.Time, ackEliciting bool) error
 	DropPackets(protocol.EncryptionLevel)
 
-	GetAlarmTimeout() time.Time
-	GetAckFrame(_ protocol.EncryptionLevel, now time.Time, onlyIfQueued bool) *wire.AckFrame
+	GetAlarmTimeout() monotime.Time
+	GetAckFrame(_ protocol.EncryptionLevel, now monotime.Time, onlyIfQueued bool) *wire.AckFrame
 }