--- a/internal/congestion/clock.go
+++ b/internal/congestion/clock.go
@@ -1,10 +1,12 @@
 package congestion
 
-import "time"
+import (
+	"github.com/quic-go/quic-go/internal/monotime"
+)
 
 // A Clock returns the current time
 type Clock interface {
-	Now() time.Time
+	Now() monotime.Time
 }
 
 // DefaultClock implements the Clock interface using the Go stdlib clock.
@@ -13,6 +15,6 @@ type DefaultClock struct{}
 var _ Clock = DefaultClock{}
 
 // Now gets the current time
-func (DefaultClock) Now() time.Time {
-	return time.Now()
+func (DefaultClock) Now() monotime.Time {
+	return monotime.Now()
 }