We need to do a lot of timer computations when running a QUIC connection. In general, for every packet sent, we need to (re)set the loss detection timer, and similarly, for (almost) every packet received, we need to set a timer to send and acknowledgment for that packet.

This PR implements a monotonic drop-in replacement for `time.Time` that uses a single `int64`. For the purposes of internals of a QUIC connections, we don't care about the wall clock time, all we're interested are accurate time differences, i.e. monotonic clock readings.

In addition to the timer calculations described above, we store the time a packet was sent in the packet history. This is needed so we can run loss detection. Storing a `time.Time` costs 24 bytes, whereas an `int64` just consumes 8 bytes. This might not seem like a lot, but when storing (say) 10000 packets that are in flight, it adds up.

I was surprised by the speedup achieved by this change:
```
name                                                   old time/op    new time/op    delta
SendAndAcknowledge/ack_every:_2,_in_flight:_0-16          112ns ± 7%      84ns ± 3%  -24.78%  (p=0.008 n=5+5)
SendAndAcknowledge/ack_every:_10,_in_flight:_100-16      86.7ns ± 2%    62.7ns ± 2%  -27.60%  (p=0.008 n=5+5)
SendAndAcknowledge/ack_every:_100,_in_flight:_1000-16    79.9ns ± 3%    58.3ns ± 3%  -27.08%  (p=0.008 n=5+5)

name                                                   old alloc/op   new alloc/op   delta
SendAndAcknowledge/ack_every:_2,_in_flight:_0-16          8.00B ± 0%     8.00B ± 0%     ~     (all equal)
SendAndAcknowledge/ack_every:_10,_in_flight:_100-16       14.0B ± 0%     14.0B ± 0%     ~     (all equal)
SendAndAcknowledge/ack_every:_100,_in_flight:_1000-16     24.0B ± 0%     24.0B ± 0%     ~     (all equal)

name                                                   old allocs/op  new allocs/op  delta
SendAndAcknowledge/ack_every:_2,_in_flight:_0-16           0.00           0.00          ~     (all equal)
SendAndAcknowledge/ack_every:_10,_in_flight:_100-16        0.00           0.00          ~     (all equal)
SendAndAcknowledge/ack_every:_100,_in_flight:_1000-16      0.00           0.00          ~     (all equal)
```